+++ core/components/romanescobackyard/model/romanescobackyard/metadata.mysql.php
@@ -12,4 +12,14 @@ $xpdo_meta_map = array (
     6 => 'rmOption',
     7 => 'rmOptionGroup',
   ),
+  'rmTask' => 
+  array (
+    0 => 'rmTaskResource',
+  ),
+  'rmSocialConnect' => 
+  array (
+    0 => 'rmSocialConnectGlobal',
+    1 => 'rmSocialConnectResource',
+    2 => 'rmSocialConnectUser',
+  ),
 );