+++ core/components/romanescobackyard/model/romanescobackyard/mysql/rmsocialconnectglobal.class.php
@@ -1,0 +1,7 @@
+<?php
+/**
+ * @package romanescobackyard
+ */
+require_once (strtr(realpath(dirname(dirname(__FILE__))), '\\', '/') . '/rmsocialconnectglobal.class.php');
+class rmSocialConnectGlobal_mysql extends rmSocialConnectGlobal {}
+?>