+++ core/components/romanescobackyard/model/romanescobackyard/mysql/rmsocialconnectglobal.map.inc.php
@@ -1,0 +1,19 @@
+<?php
+/**
+ * @package romanescobackyard
+ */
+$xpdo_meta_map['rmSocialConnectGlobal']= array (
+  'package' => 'romanescobackyard',
+  'version' => '1.1',
+  'extends' => 'rmSocialConnect',
+  'tableMeta' => 
+  array (
+    'engine' => 'InnoDB',
+  ),
+  'fields' => 
+  array (
+  ),
+  'fieldMeta' => 
+  array (
+  ),
+);