+++ core/components/romanescobackyard/model/romanescobackyard/mysql/rmsocialconnectresource.class.php
@@ -1,0 +1,7 @@
+<?php
+/**
+ * @package romanescobackyard
+ */
+require_once (strtr(realpath(dirname(dirname(__FILE__))), '\\', '/') . '/rmsocialconnectresource.class.php');
+class rmSocialConnectResource_mysql extends rmSocialConnectResource {}
+?>