+++ core/components/romanescobackyard/model/romanescobackyard/mysql/rmsocialconnectresource.map.inc.php
@@ -1,0 +1,30 @@
+<?php
+/**
+ * @package romanescobackyard
+ */
+$xpdo_meta_map['rmSocialConnectResource']= array (
+  'package' => 'romanescobackyard',
+  'version' => '1.1',
+  'extends' => 'rmSocialConnect',
+  'tableMeta' => 
+  array (
+    'engine' => 'InnoDB',
+  ),
+  'fields' => 
+  array (
+  ),
+  'fieldMeta' => 
+  array (
+  ),
+  'aggregates' => 
+  array (
+    'Resource' => 
+    array (
+      'class' => 'modResource',
+      'local' => 'parent_id',
+      'foreign' => 'id',
+      'cardinality' => 'one',
+      'owner' => 'foreign',
+    ),
+  ),
+);