+++ core/components/romanescobackyard/model/romanescobackyard/mysql/rmsocialconnectuser.class.php
@@ -1,0 +1,7 @@
+<?php
+/**
+ * @package romanescobackyard
+ */
+require_once (strtr(realpath(dirname(dirname(__FILE__))), '\\', '/') . '/rmsocialconnectuser.class.php');
+class rmSocialConnectUser_mysql extends rmSocialConnectUser {}
+?>