+++ core/components/romanescobackyard/model/romanescobackyard/rmsocialconnectresource.class.php
@@ -1,0 +1,6 @@
+<?php
+/**
+ * @package romanescobackyard
+ */
+class rmSocialConnectResource extends rmSocialConnect {}
+?>