+++ core/components/romanescobackyard/model/romanescobackyard/rmsocialconnectuser.class.php
@@ -1,0 +1,6 @@
+<?php
+/**
+ * @package romanescobackyard
+ */
+class rmSocialConnectUser extends rmSocialConnect {}
+?>