+++ core/components/romanescobackyard/docs/changelog.md
@@ -5,6 +5,7 @@ Released on ...
 
 - Add search to Tasks grid
 - Allow Tasks to be attached to objects other than resource
+- Allow social connections to be attached to objects other than resource
 - Tweak lexicon keys for Global Content top menu
 - Stop using extension package to autoload base class
 - Evaluate widest breakpoint when generating critical CSS (1920px)